% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bays29_coord}
\alias{Bays29_coord}
\title{Bays29 Coordinate Data}
\format{A collection of information about well studied TSP instances
\describe{
  \item{coord}{The coordinate matrix}
  \item{Course}{The name of the map}
  \item{Node}{A numbered list of nodes for reference}
  \item{X}{The X coordinates of the nodes}
  \item{Y}{The Y coordinates of the nodes}
  }}
\usage{
Bays29_coord
}
\description{
Dataset curated by TSPLIB
}
\keyword{datasets}
