% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bays29_dist}
\alias{Bays29_dist}
\title{Bays29 Distance Matrix}
\format{A collection of information about well studied TSP instances
\describe{
  \item{dist}{The distance matrix}
  }}
\usage{
Bays29_dist
}
\description{
Dataset curated by TSPLIB
}
\seealso{
Other Bays29: 
\code{\link{Bays29_coord}},
\code{\link{Bays29_optimal}}
}
\concept{Bays29}
\keyword{datasets}
