% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{beers_open_ended}
\alias{beers_open_ended}
\title{Beers Ordinary Graduation with Open-Ended Age Group}
\usage{
beers_open_ended(x)
}
\arguments{
\item{x}{vector of length at least 6}
}
\description{
Graduates vector from 5-year to single-year age,
assuming the final value is an open-ended age group
rather than an additional 5-year age group to be
graduated.
}
