% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Berlin52_dist}
\alias{Berlin52_dist}
\title{Berlin52 Distance Matrix}
\format{A collection of information about well studied TSP instances
\describe{
  \item{dist}{The distance matrix}
  }}
\usage{
Berlin52_dist
}
\description{
Dataset curated by TSPLIB
}
\keyword{datasets}
