% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cho130_coord}
\alias{Cho130_coord}
\title{Cho130 Coordinate Data}
\format{A collection of information about well studied TSP instances
\describe{
  \item{coord}{The coordinate matrix}
  \item{Course}{The name of the map}
  \item{Node}{A numbered list of nodes for reference}
  \item{X}{The X coordinates of the nodes}
  \item{Y}{The Y coordinates of the nodes}
  }}
\usage{
Cho130_coord
}
\description{
Dataset curated by TSPLIB
}
\seealso{
Other Cho130: 
\code{\link{Cho130_dist}},
\code{\link{Cho130_optimal}}
}
\concept{Cho130}
\keyword{datasets}
