% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cho130_dist}
\alias{Cho130_dist}
\title{Cho130 Distance Matrix}
\format{A collection of information about well studied TSP instances
\describe{
  \item{dist}{The distance matrix}
  }}
\usage{
Cho130_dist
}
\description{
Dataset curated by TSPLIB
}
\keyword{datasets}
