% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cho130_optimal}
\alias{Cho130_optimal}
\title{Cho130 Optimal Tour Length}
\format{A collection of information about well studied TSP instances
\describe{
  \item{optimal}{The optimal tour length}
  }}
\usage{
Cho130_optimal
}
\description{
Dataset curated by TSPLIB
}
\keyword{datasets}
