% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{deserialize_tensor_to_r}
\alias{deserialize_tensor_to_r}
\title{Deseralize an Eigen::Tensor into an R array}
\usage{
deserialize_tensor_to_r(path)
}
\arguments{
\item{path}{Path to serialized data}
}
\value{
An R array containing the deserialized data
}
\description{
Deseralize an Eigen::Tensor into an R array
}
\keyword{internal}
