% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eil51_coord}
\alias{eil51_coord}
\title{eil51 Coordinate Data}
\format{A collection of information about well studied TSP instances
\describe{
  \item{coord}{The coordinate matrix}
  \item{Course}{The name of the map}
  \item{Node}{A numbered list of nodes for reference}
  \item{X}{The X coordinates of the nodes}
  \item{Y}{The Y coordinates of the nodes}
  }}
\usage{
eil51_coord
}
\description{
Dataset curated by TSPLIB
}
\seealso{
Other eil51: 
\code{\link{eil51_dist}},
\code{\link{eil51_optimal}}
}
\concept{eil51}
\keyword{datasets}
