% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eil51_dist}
\alias{eil51_dist}
\title{eil51 Distance Matrix}
\format{A collection of information about well studied TSP instances
\describe{
  \item{dist}{The distance matrix}
  }}
\usage{
eil51_dist
}
\description{
Dataset curated by TSPLIB
}
\keyword{datasets}
