% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eil51_optimal}
\alias{eil51_optimal}
\title{eil51 Optimal Tour Length}
\format{A collection of information about well studied TSP instances
\describe{
  \item{optimal}{The optimal tour length}
  }}
\usage{
eil51_optimal
}
\description{
Dataset curated by TSPLIB
}
\keyword{datasets}
