% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cpp.R
\name{generate_parameter_types}
\alias{generate_parameter_types}
\title{Generate C++ for input parameter types}
\usage{
generate_parameter_types(
  dest,
  input_csv = frogger_file("cpp_generation/model_input.csv")
)
}
\arguments{
\item{dest}{The destination to write generated code to.}
}
\value{
Nothing, called to generate code in src dir
}
\description{
This generates using metadata from \code{inst/cpp_generation/model_input.csv}
}
\keyword{internal}
