% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Distance Data for the Travelling Salesman Problem}
\usage{
get_data(
  course = c("eil51", "ts225", "pr1002", "gr120", "rat195", "Bays29", "Berlin52",
    "Cho130", "KroA100", "pcb442", "pr76", "gr48", "pma343")
)
}
\arguments{
\item{course}{A \code{string} description of the desired instance information.}
}
\description{
Outputs a list containing the name, distance matrix, coordinate matrix, and known optimal value from a TSPLIB data instance
}
\examples{
put example here
}
