% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cpp.R
\name{get_data_by_struct}
\alias{get_data_by_struct}
\title{Organise the data into separate structs}
\usage{
get_data_by_struct(data)
}
\arguments{
\item{data}{Data related to model inputs or outputs}
}
\description{
This takes a list of inputs or outputs and splits them by struct returning
the result as a named list of lists where names are the struct name and
list is the inputs/outputs which belong on that struct
}
\keyword{internal}
