% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{get_time_slice}
\alias{get_time_slice}
\title{Slice a single year from model state}
\usage{
get_time_slice(state, index)
}
\arguments{
\item{state}{The model state with time dimension}

\item{index}{The index of the time step you want to extract}
}
\value{
List of model outputs for the specified time step. Can be used as
input state for \code{\link[=run_model_from_state]{run_model_from_state()}} and \code{\link[=run_model_single_year]{run_model_single_year()}}.
All outputs will have 1 fewer dimension than input state.
}
\description{
Slice a single year from model state
}
