% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gr120_coord}
\alias{gr120_coord}
\title{gr120 Coordinate Data}
\format{A collection of information about well studied TSP instances
\describe{
  \item{coord}{The coordinate matrix}
  \item{Course}{The name of the map}
  \item{Node}{A numbered list of nodes for reference}
  \item{X}{The X coordinates of the nodes}
  \item{Y}{The Y coordinates of the nodes}
  }}
\usage{
gr120_coord
}
\description{
Dataset curated by TSPLIB
}
\seealso{
Other gr120: 
\code{\link{gr120_dist}},
\code{\link{gr120_optimal}}
}
\concept{gr120}
\keyword{datasets}
