% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gr120_dist}
\alias{gr120_dist}
\title{gr120 Distance Matrix}
\format{A collection of information about well studied TSP instances
\describe{
  \item{dist}{The distance matrix}
  }}
\usage{
gr120_dist
}
\description{
Dataset curated by TSPLIB
}
\seealso{
Other gr120: 
\code{\link{gr120_coord}},
\code{\link{gr120_optimal}}
}
\concept{gr120}
\keyword{datasets}
