% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gr120_optimal}
\alias{gr120_optimal}
\title{gr120 Optimal Tour Length}
\format{A collection of information about well studied TSP instances
\describe{
  \item{optimal}{The optimal tour length}
  }}
\usage{
gr120_optimal
}
\description{
Dataset curated by TSPLIB
}
\seealso{
Other gr120: 
\code{\link{gr120_coord}},
\code{\link{gr120_dist}}
}
\concept{gr120}
\keyword{datasets}
