% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gr48_coord}
\alias{gr48_coord}
\title{gr48 Coordinate Data}
\format{A collection of information about well studied TSP instances
\describe{
  \item{coord}{The coordinate matrix}
  \item{Course}{The name of the map}
  \item{Node}{A numbered list of nodes for reference}
  \item{X}{The X coordinates of the nodes}
  \item{Y}{The Y coordinates of the nodes}
  }}
\usage{
gr48_coord
}
\description{
Dataset curated by TSPLIB
}
\seealso{
Other gr48: 
\code{\link{gr48_dist}},
\code{\link{gr48_optimal}}
}
\concept{gr48}
\keyword{datasets}
