% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gr48_dist}
\alias{gr48_dist}
\title{gr48 Distance Matrix}
\format{A collection of information about well studied TSP instances
\describe{
  \item{dist}{The distance matrix}
  }}
\usage{
gr48_dist
}
\description{
Dataset curated by TSPLIB
}
\keyword{datasets}
