% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gr48_optimal}
\alias{gr48_optimal}
\title{gr48 Optimal Tour Length}
\format{A collection of information about well studied TSP instances
\describe{
  \item{optimal}{The optimal tour length}
  }}
\usage{
gr48_optimal
}
\description{
Dataset curated by TSPLIB
}
\seealso{
Other gr48: 
\code{\link{gr48_coord}},
\code{\link{gr48_dist}}
}
\concept{gr48}
\keyword{datasets}
