% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{group_list_of_lists}
\alias{group_list_of_lists}
\title{Take a list of lists and split into groups based on some property}
\usage{
group_list_of_lists(list, on)
}
\arguments{
\item{list}{List of lists to split}

\item{on}{Name of property to split on}
}
\value{
List of lists of lists :O,
}
\description{
Every list must have this property otherwise this will fail
}
\examples{
input_list <- list(
  list(id = 1, type = "foo"),
  list(id = 2, type = "foo"),
  list(id = 3, type = "bar"),
  list(id = 4, type = "bar")
)
group_list_of_lists(input_list, "type")
}
\keyword{internal}
