% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportData.R
\name{ImportData}
\alias{ImportData}
\title{Converts raw inputs into an LF class object}
\usage{
ImportData(
  distances = NULL,
  coordinates = NULL,
  tour = NULL,
  optimal = NA,
  knownOpt = NA,
  latlon = FALSE
)
}
\arguments{
\item{distances}{A \code{n}x\code{n matrix} where first row is not column headers.}

\item{coordinates}{A \code{n}x\code{2 or 3} data object where first row is not column headers. Not required, but node names may be included as the first column.}

\item{tour}{An \code{integer vector} containing the order to visit each node in a solution.}

\item{optimal}{A \code{numeric} containing the solution tour length.}

\item{knownOpt}{A \code{numeric} containing the known optimal tour length.}

\item{latlon}{A \code{boolean} TRUE if coordinates are latitudes and longitudes, 0 otherwise.}
}
\description{
Takes a distance, coordinate, and tour information and creates an LF class object.
}
