% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KroA100_dist}
\alias{KroA100_dist}
\title{KroA100 Distance Matrix}
\format{A collection of information about well studied TSP instances
\describe{
  \item{dist}{The distance matrix}
  }}
\usage{
KroA100_dist
}
\description{
Dataset curated by TSPLIB
}
\keyword{datasets}
