% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KroA100_optimal}
\alias{KroA100_optimal}
\title{KroA100 Optimal Tour Length}
\format{A collection of information about well studied TSP instances
\describe{
  \item{optimal}{The optimal tour length}
  }}
\usage{
KroA100_optimal
}
\description{
Dataset curated by TSPLIB
}
\seealso{
Other KroA100: 
\code{\link{KroA100_coord}},
\code{\link{KroA100_dist}}
}
\concept{KroA100}
\keyword{datasets}
