% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandDist.R
\name{LandDist}
\alias{LandDist}
\title{Calculates opportunity costs to insert a node}
\usage{
LandDist(dMat, p, nodeCount, tour, jumpers, node)
}
\arguments{
\item{tour}{A \code{numeric} vector with the order that the nodes in the coordMatrix are visited. Must be the same length as the coordMatrix}

\item{distMatrix}{A \code{n}x\code{n matrix} where first row is not column headers.
Each cell represents the distance from the row index node to the column index node.}
}
\description{
Takes a tour and distance matrix and returns the tour length
}
