\name{LEAPFrOG-package}\alias{LEAPFrOG-package}\alias{LEAPFrOG-package}\docType{package}\title{Likelihood Estimation of Admixture in Parents From Offspring Genotypes}\description{Takes genotype data for a single individual, and allele frequencies for multiple populations, and returns estimations of admixture proportions for the individual, as well as the admixture proportion of their two parents.}\details{\tabular{ll}{Package: \tab LEAPFROG\crType: \tab Package\crVersion: \tab 1.0\crDate: \tab 2011-06-08\crLicense: \tab GPL\crLazyLoad: \tab yes\cr}Use LEAPFrOG when regular genotype data is available (0,1 or 2 alleles at each SNP).Use LEAPFrOG_EM when phased haplotypes are available.Plot results using LEAPFrOG_plot}\author{Daniel Crouch and Michael WealeMaintainer: Daniel Crouch - daniel.crouch@oncology.ox.ac.uk}\references{Crouch & Weale (2011), European Journal of Human Genetics}\seealso{\code{\link[LEAPFrOG:LEAPFrOG_plot]{LEAPFrOG_plot}},\code{\link[LEAPFrOG:LEAPFrOG_EM]{LEAPFrOG_EM}},\code{\link[LEAPFrOG:LEAPFrOG]{LEAPFrOG}},\code{\link[LEAPFrOG:BEAPFrOG]{BEAPFrOG}}}