% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{leapfrogR}
\alias{leapfrogR}
\title{Simulate leapfrog model}
\usage{
leapfrogR(demp, projp, hiv_strat_type = "full", hiv_steps_per_year = 10L)
}
\arguments{
\item{basepop}{base population: matrix indexed by (age, sex)}

\item{sx}{three-dimensional array of survival probabilities indexed by (age, sex, year)}

\item{asfr}{two-dimensional array of age-specific fertility rates (age, year)}
}
\description{
Simulate leapfrog model
}
\details{
The first year of `sx`, `asfr`, `srb`, and `netmig` is not used. This is assumed
to apply to the base year population (consistent with Spectrum).
}
