% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF.R
\name{LF}
\alias{LF}
\title{LeapFrog metaheuristic search algorithm}
\usage{
LF(LFObj, p = 1, m = 1, s = 10^-3, r = 1, a = 0, monitor = TRUE)
}
\arguments{
\item{LFObj}{A LeapFrog class object created by running the ImportData() function.}

\item{p}{A \code{double} (0,1] which represents the ratio of the maximum number of nodes removed in each iteration of the LF algorithm.}

\item{m}{An \code{integer} (0,inf) which represents the number of games played}

\item{s}{A \code{double} (0,1] which represents the uncertainty used in the first round of each game}

\item{r}{An \code{integer} (0,inf) which represents the number of rounds in each game}

\item{a}{A \code{double} (0,1] which represents the variable number of players used in successive rounds}

\item{monitor}{A \code{boolean} or function used to display information during execution of the algorithm}
}
\description{
Takes a distance matrix and searches for an optimal tour
}
