% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NN.R
\name{NN}
\alias{NN}
\title{Nearest Neighbor algorithm}
\usage{
NN(LFObj)
}
\arguments{
\item{LFObj}{A LeapFrog class object}
}
\description{
Runs the nearest neighbor algorithm starting from each node, then returns shortest tour
}
