% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NN.R
\name{NNCompare}
\alias{NNCompare}
\title{Nearest Neighbor comparison tool}
\usage{
NNCompare(LFObj)
}
\arguments{
\item{LFObj}{A LeapFrog class object}
}
\description{
Runs the nearest neighbor function and adds the results to the lfHist plot
}
