% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcb442_coord}
\alias{pcb442_coord}
\title{pcb442 Coordinate Data}
\format{A collection of information about well studied TSP instances
\describe{
  \item{coord}{The coordinate matrix}
  \item{Course}{The name of the map}
  \item{Node}{A numbered list of nodes for reference}
  \item{X}{The X coordinates of the nodes}
  \item{Y}{The Y coordinates of the nodes}
  }}
\usage{
pcb442_coord
}
\description{
Dataset curated by TSPLIB
}
\seealso{
Other pcb442: 
\code{\link{pcb442_dist}},
\code{\link{pcb442_optimal}}
}
\concept{pcb442}
\keyword{datasets}
