% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcb442_dist}
\alias{pcb442_dist}
\title{pcb442 Distance Matrix}
\format{A collection of information about well studied TSP instances
\describe{
  \item{dist}{The distance matrix}
  }}
\usage{
pcb442_dist
}
\description{
Dataset curated by TSPLIB
}
\seealso{
Other pcb442: 
\code{\link{pcb442_coord}},
\code{\link{pcb442_optimal}}
}
\concept{pcb442}
\keyword{datasets}
