% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcb442_optimal}
\alias{pcb442_optimal}
\title{pcb442 Optimal Tour Length}
\format{A collection of information about well studied TSP instances
\describe{
  \item{optimal}{The optimal tour length}
  }}
\usage{
pcb442_optimal
}
\description{
Dataset curated by TSPLIB
}
\seealso{
Other pcb442: 
\code{\link{pcb442_coord}},
\code{\link{pcb442_dist}}
}
\concept{pcb442}
\keyword{datasets}
