% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTour.R
\name{PlotTour}
\alias{PlotTour}
\title{Plot coordinates with tour}
\usage{
PlotTour(LFObj)
}
\arguments{
\item{LFObj}{A \code{LeapFrog} class object}
}
\description{
Takes a coordinate matrix and plots a tour through it
}
