% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pma343_coord}
\alias{pma343_coord}
\title{pma343 Coordinate Data}
\format{A collection of information about well studied TSP instances
\describe{
  \item{coord}{The coordinate matrix}
  \item{Course}{The name of the map}
  \item{Node}{A numbered list of nodes for reference}
  \item{X}{The X coordinates of the nodes}
  \item{Y}{The Y coordinates of the nodes}
  }}
\usage{
pma343_coord
}
\description{
Dataset curated by TSPLIB
}
\seealso{
Other pma343: 
\code{\link{pma343_dist}},
\code{\link{pma343_optimal}}
}
\concept{pma343}
\keyword{datasets}
