% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pma343_dist}
\alias{pma343_dist}
\title{pma343 Distance Matrix}
\format{A collection of information about well studied TSP instances
\describe{
  \item{dist}{The distance matrix}
  }}
\usage{
pma343_dist
}
\description{
Dataset curated by TSPLIB
}
\seealso{
Other pma343: 
\code{\link{pma343_coord}},
\code{\link{pma343_optimal}}
}
\concept{pma343}
\keyword{datasets}
