% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pma343_optimal}
\alias{pma343_optimal}
\title{pma343 Optimal Tour Length}
\format{A collection of information about well studied TSP instances
\describe{
  \item{optimal}{The optimal tour length}
  }}
\usage{
pma343_optimal
}
\description{
Dataset curated by TSPLIB
}
\keyword{datasets}
