% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pr1002_coord}
\alias{pr1002_coord}
\title{pr1002 Coordinate Data}
\format{A collection of information about well studied TSP instances
\describe{
  \item{coord}{The coordinate matrix}
  \item{Course}{The name of the map}
  \item{Node}{A numbered list of nodes for reference}
  \item{X}{The X coordinates of the nodes}
  \item{Y}{The Y coordinates of the nodes}
  }}
\usage{
pr1002_coord
}
\description{
Dataset curated by TSPLIB
}
\seealso{
Other pr1002: 
\code{\link{pr1002_dist}},
\code{\link{pr1002_optimal}}
}
\concept{pr1002}
\keyword{datasets}
