% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pr76_optimal}
\alias{pr76_optimal}
\title{pr76 Optimal Tour Length}
\format{A collection of information about well studied TSP instances
\describe{
  \item{optimal}{The optimal tour length}
  }}
\usage{
pr76_optimal
}
\description{
Dataset curated by TSPLIB
}
\keyword{datasets}
