% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_inputs.R
\name{prepare_leapfrog_demp}
\alias{prepare_leapfrog_demp}
\title{Prepare demographic inputs from Spectrum PJNZ}
\usage{
prepare_leapfrog_demp(pjnz)
}
\arguments{
\item{pjnz}{path to PJNZ file}
}
\value{
list of demographic input parameters
}
\description{
Prepare demographic inputs from Spectrum PJNZ
}
\examples{
pjnz <- system.file(
  "pjnz/bwa_aim-adult-art-no-special-elig_v6.13_2022-04-18.PJNZ",
  package = "frogger")
demp <- prepare_leapfrog_demp(pjnz)

}
