% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_inputs.R
\name{prepare_leapfrog_parameters}
\alias{prepare_leapfrog_parameters}
\title{Prepare leapfrog input parameters from Spectrum PJNZ}
\usage{
prepare_leapfrog_parameters(pjnz)
}
\arguments{
\item{pjnz}{path to PJNZ file}
}
\value{
list of demographic and HIV projection input parameters
}
\description{
Prepare leapfrog input parameters from Spectrum PJNZ
}
\examples{
pjnz <- system.file(
  "pjnz/bwa_aim-adult-art-no-special-elig_v6.13_2022-04-18.PJNZ",
  package = "frogger")
parameters <- prepare_leapfrog_parameters(pjnz)

}
