% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum-inputs.R
\name{prepare_leapfrog_projp}
\alias{prepare_leapfrog_projp}
\title{Prepare adult HIV projection parameters from Spectrum PJNZ}
\usage{
prepare_leapfrog_projp(pjnz, hiv_steps_per_year = 10L, hTS = 3)
}
\description{
Prepare adult HIV projection parameters from Spectrum PJNZ
}
