% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{process_parameters_to_cpp}
\alias{process_parameters_to_cpp}
\title{Process parameters and convert from 1 based indexing in R to
0 based indexing in C++. Also add in any defaults/extra parameters,
e.g. \code{h_art_stage_dur}.}
\usage{
process_parameters_to_cpp(parameters)
}
\arguments{
\item{parameters}{The list of parameters to feed into the model.}
}
\value{
List of parameters with 0 based indexing and defaults.
}
\description{
Process parameters and convert from 1 based indexing in R to
0 based indexing in C++. Also add in any defaults/extra parameters,
e.g. \code{h_art_stage_dur}.
}
