% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{process_parameters_to_r}
\alias{process_parameters_to_r}
\title{Process parameters and convert from 0 based indexing for C++ to
1 based indexing in R.}
\usage{
process_parameters_to_r(parameters)
}
\arguments{
\item{parameters}{List of parameters.}
}
\value{
List of parameters with 1 based indexing.
}
\description{
Process parameters and convert from 0 based indexing for C++ to
1 based indexing in R.
}
