% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rat195_dist}
\alias{rat195_dist}
\title{rat195 Distance Matrix}
\format{A collection of information about well studied TSP instances
\describe{
  \item{dist}{The distance matrix}
  }}
\usage{
rat195_dist
}
\description{
Dataset curated by TSPLIB
}
\keyword{datasets}
