% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rat195_optimal}
\alias{rat195_optimal}
\title{rat195 Optimal Tour Length}
\format{A collection of information about well studied TSP instances
\describe{
  \item{optimal}{The optimal tour length}
  }}
\usage{
rat195_optimal
}
\description{
Dataset curated by TSPLIB
}
\keyword{datasets}
