% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-spectrum.R
\name{read_dp}
\alias{read_dp}
\title{Read Spectrum .DP file}
\usage{
read_dp(pjnz)
}
\arguments{
\item{pjnz}{file path to Spectrum PJNZ file}
}
\value{
Matrix with class "spectrum_dp". Not intended for direct use.
}
\description{
This function returns the Spectrum .DP file read as character CSV. Not intended
for direct use, but passing to other functions to parse.
}
\examples{

pjnz <- system.file("pjnz/bwa2021_v6.13.pjnz", package = "leapfrog")
dp <- read_dp(pjnz)
class(dp)

}
