% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5_utils.R
\name{read_hdf5_file}
\alias{read_hdf5_file}
\title{This function reads an HDF5 file to an R dataframe.}
\usage{
read_hdf5_file(file_path)
}
\arguments{
\item{file_path}{where the HDF5 file is}
}
\value{
deserialized HDF5 file as a list
}
\description{
This function reads an HDF5 file to an R dataframe.
}
