% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5_utils.R
\name{read_parameters}
\alias{read_parameters}
\title{Read parameters from HDF5 file format. This implicitly
processes the parameters from C++ 0 based indexing to
R 1 based indexing.}
\usage{
read_parameters(file_path)
}
\arguments{
\item{file_path}{HDF5 file to read}
}
\description{
Read parameters from HDF5 file format. This implicitly
processes the parameters from C++ 0 based indexing to
R 1 based indexing.
}
