% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model_single_year}
\alias{run_model_single_year}
\title{Run leapfrog model fit for a single year}
\usage{
run_model_single_year(
  parameters,
  configuration,
  initial_state,
  start_from_year
)
}
\arguments{
\item{parameters}{Projection parameters}

\item{configuration}{The model configuration to run, see
\code{\link[=list_model_configurations]{list_model_configurations()}} for available configurations}

\item{initial_state}{The model will run from this initial state}

\item{start_from_year}{Start the model simulation from this year}
}
\value{
List of model outputs
}
\description{
Run leapfrog model fit for a single year
}
