% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_wlhiv_births}
\alias{run_wlhiv_births}
\title{Run run_wlhiv_births c++ function without running adult model}
\usage{
run_wlhiv_births(
  data,
  parameters,
  sim_years,
  hts_per_year,
  data_from_adult_model,
  output_steps = NULL
)
}
\arguments{
\item{data}{Input data}

\item{parameters}{Projection parameters}

\item{sim_years}{Simulation years to run model for default 1970:2030}

\item{hts_per_year}{Number of HIV time steps per year, default 10}

\item{data_from_adult_model}{This is a list with parameters h_hiv_adult,
h_art_adult, births, p_total_pop from a previous full model run.}

\item{output_steps}{Which sim years to output for default same as sim_years}
}
\value{
List of model outputs
}
\description{
Run run_wlhiv_births c++ function without running adult model
}
