% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA.R
\name{SACompare}
\alias{SACompare}
\title{Simulated Annealing Comparison}
\usage{
SACompare(LFObj, temp = 200, alpha = 0.9)
}
\arguments{
\item{LFObj}{A LeapFrog class object}

\item{temp}{The starting temperature}

\item{alpha}{The annealing rate}
}
\description{
Runs the Simulated Annealing Algorithm for the same time duration that Leap Frog required
}
