% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5_utils.R
\name{save_hdf5_file}
\alias{save_hdf5_file}
\title{This function saves an arbitrary list/dataframe to HDF5
file format. Note that this file format can be used
in any language we support which means some R specific
features of the data are not guaranteed to be preserved,
e.g. dimnames}
\usage{
save_hdf5_file(df, file_path)
}
\arguments{
\item{df}{list/dataframe to serialize}

\item{file_path}{where to save the HDF5 file}
}
\description{
This function saves an arbitrary list/dataframe to HDF5
file format. Note that this file format can be used
in any language we support which means some R specific
features of the data are not guaranteed to be preserved,
e.g. dimnames
}
