% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5_utils.R
\name{save_parameters}
\alias{save_parameters}
\title{Save parameters to HDF5 file format. This implicitly
processes the parameters to C++ 0 based indexing.}
\usage{
save_parameters(df, file_path)
}
\arguments{
\item{df}{list/dataframe to serialize}

\item{file_path}{where to save the HDF5 file}
}
\description{
Save parameters to HDF5 file format. This implicitly
processes the parameters to C++ 0 based indexing.
}
