% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{serialize_r_to_tensor}
\alias{serialize_r_to_tensor}
\title{Serialize an R array into a format that can be read as an Eigen::Tensor
by deserialize_tensor}
\usage{
serialize_r_to_tensor(data, path)
}
\arguments{
\item{data}{The data to serialize, should be an array or vector}

\item{path}{File path to serialize data to}
}
\value{
Path to file
}
\description{
Serialize an R array into a format that can be read as an Eigen::Tensor
by deserialize_tensor
}
\keyword{internal}
