% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tour_length.R
\name{tour_length}
\alias{tour_length}
\title{Calculate tour length}
\usage{
tour_length(distMatrix, tour)
}
\arguments{
\item{distMatrix}{A \code{n}x\code{n matrix} where first row is not column headers.
Each cell represents the distance from the row index node to the column index node.}

\item{tour}{A \code{numeric} vector with the order that the nodes in the coordMatrix are visited. Must be the same length as the coordMatrix}
}
\description{
Takes a tour and distance matrix and returns the tour length
}
\examples{
put example here

}
