% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts225_coord}
\alias{ts225_coord}
\title{ts225 Coordinate Data}
\format{A collection of information about well studied TSP instances
\describe{
  \item{coord}{The coordinate matrix}
  \item{Course}{The name of the map}
  \item{Node}{A numbered list of nodes for reference}
  \item{X}{The X coordinates of the nodes}
  \item{Y}{The Y coordinates of the nodes}
  }}
\usage{
ts225_coord
}
\description{
Dataset curated by TSPLIB
}
\seealso{
Other ts225: 
\code{\link{ts225_dist}},
\code{\link{ts225_optimal}}
}
\concept{ts225}
\keyword{datasets}
