% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts225_dist}
\alias{ts225_dist}
\title{ts225 Distance Matrix}
\format{A collection of information about well studied TSP instances
\describe{
  \item{dist}{The distance matrix}
  }}
\usage{
ts225_dist
}
\description{
Dataset curated by TSPLIB
}
\keyword{datasets}
