% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bays29_optimal}
\alias{Bays29_optimal}
\title{Bays29 Optimal Tour Length}
\format{A collection of information about well studied TSP instances
\describe{
  \item{optimal}{The optimal tour length}
  }}
\usage{
Bays29_optimal
}
\description{
Dataset curated by TSPLIB
}
\keyword{datasets}
