% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Berlin52_optimal}
\alias{Berlin52_optimal}
\title{Berlin52 Optimal Tour Length}
\format{A collection of information about well studied TSP instances
\describe{
  \item{optimal}{The optimal tour length}
  }}
\usage{
Berlin52_optimal
}
\description{
Dataset curated by TSPLIB
}
\keyword{datasets}
