% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D2C.R
\name{D2C}
\alias{D2C}
\title{distMatrix to coordDF}
\usage{
D2C(distMatrix)
}
\arguments{
\item{distMatrix}{A \code{n}x\code{n matrix} where first row is not column headers.
Each cell represents the distance from the row index node to the column index node.}
}
\description{
Takes a distance matrix and converts it into a coordinate tibble
}
