% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{leapfrogR}
\alias{leapfrogR}
\title{Simulate leapfrog model}
\usage{
leapfrogR(demp, projp, hiv_strat = "full", hiv_steps_per_year = 10L)
}
\arguments{
\item{demp}{list of demographic input parameters (TODO: document)}

\item{projp}{list of HIV projection parameters (TODO: document)}

\item{hiv_strat}{stratification of HIV population, either "full"
(default; single-year ages) or "coarse" (aggregated age groups).}

\item{hiv_steps_per_year}{number of Euler integration steps per year
for HIV progression; default 10.}
}
\description{
Simulate leapfrog model
}
\details{
The first year of `sx`, `asfr`, `srb`, and `netmig` is not used. This is assumed
to apply to the base year population (consistent with Spectrum).
}
