% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{leapfrogR}
\alias{leapfrogR}
\title{Simulate leapfrog model}
\usage{
leapfrogR(demp)
}
\arguments{
\item{demp}{list of demographic input parameters (TODO: document)}
}
\value{
a list containing projection output arrays

 * `totpop1`: Total population by single-age, sex, and year.
 * `natdeaths`: Non-AIDS deaths to total population by single-age, sex, and year.
}
\description{
Simulate leapfrog model
}
\details{
The first year of `sx`, `asfr`, `srb`, and `netmig` is not used. This is assumed
to apply to the base year population (consistent with Spectrum).
}
