% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTour.R
\name{PlotTour}
\alias{PlotTour}
\title{Plot coordinates with tour}
\usage{
PlotTour(LFObj)
}
\arguments{
\item{coordDF}{A \code{n}x\code{2 or 3} data.frame or tibble where first row is not column headers. 
The first and second columns must contain the \code{x} and \code{y} coordinates of each node.
The third column (optional) must contain ids for each node.}

\item{tour}{A \code{numeric} vector with the order that the nodes in the coordDF are visited. Must be the same length as the coordDF}

\item{latlong}{A \code{boolean} value. \code{TRUE} if coordinates are latitudes and longitudes, \code{FALSE} otherwise.}
}
\description{
Takes a coordinate matrix and plots a tour through it
}
