% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pr1002_dist}
\alias{pr1002_dist}
\title{pr1002 Distance Matrix}
\format{A collection of information about well studied TSP instances
\describe{
  \item{dist}{The distance matrix}
  }}
\usage{
pr1002_dist
}
\description{
Dataset curated by TSPLIB
}
\keyword{datasets}
