% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pr1002_optimal}
\alias{pr1002_optimal}
\title{pr1002 Optimal Tour Length}
\format{A collection of information about well studied TSP instances
\describe{
  \item{optimal}{The optimal tour length}
  }}
\usage{
pr1002_optimal
}
\description{
Dataset curated by TSPLIB
}
\keyword{datasets}
