% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum-inputs.R
\name{prepare_leapfrog_demp}
\alias{prepare_leapfrog_demp}
\title{Prepare demographic inputs from Spectrum PJNZ}
\usage{
prepare_leapfrog_demp(pjnz)
}
\arguments{
\item{pjnz}{path to PJNZ file}
}
\value{
list of demographic input parameters
}
\description{
Prepare demographic inputs from Spectrum PJNZ
}
