% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Run leapfrog model fit}
\usage{
run_model(
  data,
  parameters,
  sim_years,
  hts_per_year,
  output_steps = NULL,
  hiv_age_stratification = "full",
  run_child_model = TRUE
)
}
\arguments{
\item{data}{Input data}

\item{parameters}{Projection parameters}

\item{sim_years}{Simulation years to run model for default 1970:2030}

\item{hts_per_year}{Number of HIV time steps per year, default 10}

\item{output_steps}{Which sim years to output for default same as sim_years}

\item{hiv_age_stratification}{The age stratification for HIV population,
"coarse" or "full"}

\item{run_child_model}{If TRUE then run the child model}
}
\value{
List of model outputs
}
\description{
Run leapfrog model fit
}
