% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Run leapfrog model fit}
\usage{
run_model(
  parameters,
  configuration = "HivFullAgeStratification",
  output_years = seq(1970, 2030)
)
}
\arguments{
\item{parameters}{Projection parameters}

\item{output_years}{Which years of the model to return from the simulation,
defaults to all years from 1970 to 2030. Also used to control what years
the simulation is run for. If output only 2030, simulation will be run
from}

\item{configuation}{The model configuration to run, see TODO}
}
\value{
List of model outputs
}
\description{
Run leapfrog model fit
}
