% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts225_optimal}
\alias{ts225_optimal}
\title{ts225 Optimal Tour Length}
\format{A collection of information about well studied TSP instances
\describe{
  \item{optimal}{The optimal tour length}
  }}
\usage{
ts225_optimal
}
\description{
Dataset curated by TSPLIB
}
\seealso{
Other ts225: 
\code{\link{ts225_coord}},
\code{\link{ts225_dist}}
}
\concept{ts225}
\keyword{datasets}
